Move testDropTable to BaseConnectorTest
