+++ plugin/trino-kudu/src/test/java/io/trino/plugin/kudu/AbstractKuduConnectorTest.java
@@ -276,6 +276,14 @@ public abstract class AbstractKuduConnectorTest
     }
 
     @Override
+    public void testDropTable()
+    {
+        assertThatThrownBy(super::testDropTable)
+                .hasMessage("Table partitioning must be specified using setRangePartitionColumns or addHashPartitions");
+        throw new SkipException("TODO Enable the test once Kudu connector can create tables with default partitions");
+    }
+
+    @Override
     protected String tableDefinitionForAddColumn()
     {
         return "(x VARCHAR WITH (primary_key=true)) WITH (partition_by_hash_columns = ARRAY['x'], partition_by_hash_buckets = 2)";
